requirements.txt
asgiref==3.9.2
attrs==25.3.0
cffi==2.0.0
cryptography==46.0.2
Django==5.2.6
django-admin-sortable2==2.2.8
django-appconf==1.1.0
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-cryptography==1.1
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.17
django-mptt==0.18.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
Faker==37.8.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.3.0
pycparser==2.23
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
